Move getFie() to iframe-helper.js
