+++ extensions/amp-gwd-animation/0.1/amp-gwd-animation.js
@@ -20,12 +20,13 @@ import {
   GWD_TIMELINE_EVENT,
 } from './amp-gwd-animation-impl';
 import {CSS} from '../../../build/amp-gwd-animation-0.1.css';
-import {Services, getFriendlyIframeEmbedOptional} from '../../../src/services';
+import {Services} from '../../../src/services';
 import {getDetail} from '../../../src/event-helper';
 import {
   getExistingServiceForDocInEmbedScope,
   getParentWindowFrameElement,
 } from '../../../src/service';
+import {getFriendlyIframeEmbedOptional} from '../../../src/iframe-helper';
 import {userAssert} from '../../../src/log';
 
 /**
